Split editor trace helpers
